mounter: Fix readblock() ID check
This code was copied from checksum() but should not use i here, because
in readblock() i refers to the number of tries to read this block. The
ID should not live in a different place depending on the number of tries.
In practice we'd hopefully never hit the case where this is broken.